# Conflicts:
#   frontend/web/pubspec.yaml

+++ frontend/android/app/src/main/java/com/ssafy/hifes/data/model/ARItemDto.kt
@@ -4,8 +4,8 @@ data class ARItemDto(
     var itemId: Int,
     var festivalId: Int,
     var hostId: Int,
-    var ARLatitude: Float,
-    var ARLongitude: Float,
+    var ARLatitude: Double,
+    var ARLongitude: Double,
     var ARImage: String,
     var giftInfo: String
 )
